refac: make ProposerSigner util
