+++ validity/Cargo.toml
@@ -24,6 +24,7 @@ op-succinct-client-utils.workspace = true
 op-succinct-elfs.workspace = true
 op-succinct-host-utils.workspace = true
 op-succinct-proof-utils.workspace = true
+op-succinct-proposer-utils.workspace = true
 
 # sp1
 sp1-sdk.workspace = true
@@ -59,7 +60,6 @@ futures-util = "0.3.31"
 clap.workspace = true
 tikv-jemallocator = "0.6.0"
 rustls = "0.23.23"
-url = "2.5.4"
 strum = { workspace = true, features = ["derive"] }
 strum_macros.workspace = true
